tidy: move TargetTuple to a new file
